+++ packages/apollo-client/CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### vNEXT
 - include `optimisticResponse` in the context passed to apollo-link for mutations [PR#2704](https://github.com/apollographql/apollo-client/pull/2704)
+- Error when invalid `cache-and-network` is provided as `query.fetchPolicy` within `defaultOptions`
 
 ### 2.1.1
 - fix eslint usage by fixing jsnext:main path