+++ packages/apollo-client/CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ### vNEXT
 - include `optimisticResponse` in the context passed to apollo-link for mutations [PR#2704](https://github.com/apollographql/apollo-client/pull/2704)
+- Allow optional parameter to include queries in standby mode when refetching observed queries [PR#2804](https://github.com/apollographql/apollo-client/pull/2804)
+
 
 ### 2.1.1
 - fix eslint usage by fixing jsnext:main path